requirements.txt
wheel
asgiref==3.4.1
Django==3.2.5
pytz==2021.3
sqlparse==0.4.2
requests==2.26.0
django-environ==0.4.5
django-cors-headers==3.7.0
redis==4.0.1
channels==3.0.4
channels-redis==3.3.1
djangorestframework==3.12.4
Pillow==8.4.0
django-rq==2.5.1
python-dateutil==2.8.2
rq-scheduler==0.11.0
django-filter==21.1
whitenoise==5.3.0
gunicorn==20.1.0
opencv-python==4.5.5.62
python-magic==0.4.25
coverage==6.4.1
boto3==1.24.22
django-sendfile==0.3.11
fontawesomefree==6.1.2 
validate-email-address==1
uuid==1.30
pandas
paho-mqtt==1.6.1
mysqlclient